+++ sdk/src/test/java/com/mapbox/maps/attribution/MapAttributionDelegateImplTest.kt
@@ -64,7 +64,8 @@ class MapAttributionDelegateImplTest {
       AttributionParserConfig(
         withImproveMap = false,
         withTelemetryAttribution = false,
-        withMapboxAttribution = false
+        withMapboxAttribution = false,
+        withMapboxPrivacyPolicy = false,
       )
     )
     assertEquals(1, attributions.size)
@@ -82,7 +83,8 @@ class MapAttributionDelegateImplTest {
       AttributionParserConfig(
         withImproveMap = true,
         withTelemetryAttribution = false,
-        withMapboxAttribution = false
+        withMapboxAttribution = false,
+        withMapboxPrivacyPolicy = false,
       )
     )
     assertEquals(2, attributions.size)
@@ -100,7 +102,8 @@ class MapAttributionDelegateImplTest {
       AttributionParserConfig(
         withImproveMap = false,
         withTelemetryAttribution = true,
-        withMapboxAttribution = false
+        withMapboxAttribution = false,
+        withMapboxPrivacyPolicy = false,
       )
     )
 
@@ -119,7 +122,8 @@ class MapAttributionDelegateImplTest {
       AttributionParserConfig(
         withImproveMap = false,
         withTelemetryAttribution = false,
-        withMapboxAttribution = true
+        withMapboxAttribution = true,
+        withMapboxPrivacyPolicy = false,
       )
     )
 
@@ -133,7 +137,7 @@ class MapAttributionDelegateImplTest {
     )
 
     attributions = mapAttributionDelegate.parseAttributions(context, AttributionParserConfig())
-    assertEquals(4, attributions.size)
+    assertEquals(5, attributions.size)
     var counter = 0
     for ((title, url) in attributions) {
       when (counter) {
@@ -171,6 +175,16 @@ class MapAttributionDelegateImplTest {
           )
           assertEquals(
             "Telemetry title should match", "Telemetry Settings",
+            title
+          )
+        }
+        4 -> {
+          assertEquals(
+            "Telemetry URL should match", "https://www.mapbox.com/legal/privacy#product-privacy-policy/",
+            url
+          )
+          assertEquals(
+            "Telemetry title should match", "Mapbox Privacy Policy",
             title
           )
         }